Extract LayerStackEvaluator and rename display types for clarity
Architectural refactor separating evaluation from hardware output:

- Extract LayerStackEvaluator: standalone stateful class owning layer
  evaluation (WhileTrue/OnChange/Expression/Constant), formatting,
  constant cycling, and per-layer state. No hardware dependencies.
- Extract SegmentRendering: static rendering pipeline utilities
  (AlignText, TruncateTo3, EncodeText) shared by hardware and UI.
- Rename DisplayEncoder -> SegmentEncoder (col01 protocol encoder)
- Rename FanatecDisplayManager -> SegmentDisplayController (orchestrates
  evaluator + rendering pipeline + hardware output)
- ScreenSettingsPanel now uses its own LayerStackEvaluator for preview,
  eliminating duplicated evaluation logic. The "no hardware" preview
  now evaluates the full layer stack including overlays, not just
  Constant layers.
- Add 13 new tests for LayerStackEvaluator (evaluation, IsTruthy,
  FormatValue, EvaluateLayer)

This separation enables future ITM display support: ITM devices in
legacy mode can reuse LayerStackEvaluator with a different output stage.
