+++ FanaBridge/FanatecPlugin.cs
@@ -32,7 +32,7 @@ namespace FanaBridge
         private FanatecTuningController _tuning;
         private LedEncoder _leds;
         private LegacyLedEncoder _legacyLeds;
-        private DisplayEncoder _display;
+        private SegmentEncoder _segmentEncoder;
         private NCalcEngineBase _nCalcEngine;
 
         /// <summary>Shared NCalc engine for evaluating display layer expressions.</summary>
@@ -72,8 +72,8 @@ namespace FanaBridge
         /// <summary>Shared legacy LED encoder (col01) — used by DeviceInstance LED drivers for legacy/RevStripe wheels.</summary>
         public LegacyLedEncoder LegacyLeds => _legacyLeds;
 
-        /// <summary>Shared display encoder — used by DeviceInstance display managers and wizard.</summary>
-        public DisplayEncoder Display => _display;
+        /// <summary>Shared 7-segment encoder (col01) — used by SegmentDisplayController.</summary>
+        public SegmentEncoder SegmentEncoder => _segmentEncoder;
 
         /// <summary>Shared tuning controller — used by TuningSettingsPanel for encoder config.</summary>
         public FanatecTuningController Tuning => _tuning;
@@ -100,7 +100,7 @@ namespace FanaBridge
             _device = new FanatecDevice();
             _leds = new LedEncoder(_device);
             _legacyLeds = new LegacyLedEncoder(_device);
-            _display = new DisplayEncoder(_device);
+            _segmentEncoder = new SegmentEncoder(_device);
             _tuning = new FanatecTuningController(
                 _device,
                 msg => SimHub.Logging.Current.Warn(msg),
@@ -192,7 +192,7 @@ namespace FanaBridge
             {
                 try
                 {
-                    _display.ClearDisplay();
+                    _segmentEncoder.ClearDisplay();
                 }
                 catch (Exception ex)
                 {
